feat: Move getBackgroundColorResBasedOnId to new Core:Coil module
