+++ UserAvatar/src/main/java/com/infomaniak/core/useravatar/exposed/UserAvatar.kt
@@ -45,11 +45,11 @@ import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
 import coil3.request.ImageRequest
 import coil3.request.crossfade
+import com.infomaniak.core.coil.getBackgroundColorResBasedOnId
 import com.infomaniak.core.useravatar.AvatarData
 import com.infomaniak.core.useravatar.AvatarDisplayState
 import com.infomaniak.core.useravatar.component.InitialsTextAvatar
 import com.infomaniak.core.useravatar.component.UnknownUserIcon
-import com.infomaniak.core.useravatar.extensions.getBackgroundColorResBasedOnId
 
 private val minAvatarSize = 32.dp
